ong_api/requirements.prod.txt
django==1.11
psycopg2==2.7
djangorestframework==3.6.1
uwsgi
django-cors-headers==1.2.0
pytz
djangorestframework-csv
